Add Assignment 2 files via upload
Assignment 2:

Study the HTML structure of this file and began to think about how you might parse it to extract the relevant data for each meeting. Using this knowledge about its structure, write a program in Node.js that will print to the console the street address for every meeting in your text file.

Update your GitHub repository with the relevant file(s). In Canvas, submit the URL of the specific location of this work within your data-structures GitHub repository. In a README file in the specific location of your work, include a screenshot of the (partial) output to the console.